# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Yoobe V3 API - Sistema de Orçamentos e Replicação
  description: |
    API para gerenciamento de orçamentos e replicação de produtos base.
    
    ## Fluxo de Orçamento → Aprovação → Replicação
    
    1. **Gestor** cria orçamento com produtos desejados
    2. **Admin Global** aprova ou rejeita o orçamento
    3. **Gestor** pode replicar produtos apenas após aprovação
    4. **Gestor** gerencia produtos replicados (ativar/inativar, editar)
    
    ## Autenticação
    
    Todas as rotas requerem autenticação via Bearer Token.
    - Role `admin`: Acesso total ao sistema
    - Role `manager`: Acesso limitado às funcionalidades de gestor
    - Role `user`: Acesso apenas a funcionalidades básicas
    
  version: 2.1.0
  contact:
    name: Yoobe Team
    email: support@yoobe.co
servers:
  - url: http://localhost:3000
    description: Servidor de Desenvolvimento
  - url: https://api.yoobe.co
    description: Servidor de Produção
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único do orçamento
        title:
          type: string
          description: Título do orçamento
        description:
          type: string
          description: Descrição detalhada
        status:
          type: string
          enum: [pending, approved, rejected]
          description: Status atual do orçamento
        total_amount:
          type: number
          format: float
          description: Valor total do orçamento
        company_id:
          type: string
          format: uuid
          description: ID da empresa do gestor
        manager_id:
          type: string
          format: uuid
          description: ID do gestor
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        reviewed_by:
          type: string
          format: uuid
          nullable: true
        admin_notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetItem'
    BudgetItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
        base_product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        custom_price:
          type: number
          format: float
        notes:
          type: string
          nullable: true
        base_product:
          $ref: '#/components/schemas/BaseProduct'
    BaseProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        base_price:
          type: number
          format: float
        base_points_cost:
          type: integer
        image_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        category_id:
          type: string
          format: uuid
          nullable: true
...
paths:
...
  /api/admin/orcamentos:
    get:
      summary: Listar todos os orçamentos
      description: Retorna todos os orçamentos recebidos (apenas Admin Global)
      tags: [Admin - Orçamentos]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
          description: Filtrar por status
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por empresa
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de orçamentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  budgets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        '401':
          description: Não autorizado
        '403':
          description: Acesso negado - Apenas Admin Global
  /api/admin/orcamentos/{id}/approve:
    post:
      summary: Aprovar ou rejeitar orçamento
      description: Aprova ou rejeita um orçamento (apenas Admin Global)
      tags: [Admin - Orçamentos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - admin_notes
              properties:
                action:
                  type: string
                  enum: [approve, reject]
                  description: Ação a ser executada
                admin_notes:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: Observações obrigatórias do admin
      responses:
        '200':
          description: Orçamento processado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  budget:
                    $ref: '#/components/schemas/Budget'
                  message:
                    type: string
        '400':
          description: Dados inválidos ou orçamento não está pendente
        '401':
          description: Não autorizado
        '403':
          description: Acesso negado - Apenas Admin Global
        '404':
          description: Orçamento não encontrado
...